Cleanups

+++ src/parts/flasher.cpp
@@ -1243,7 +1243,7 @@ void Flasher::Render(const unsigned int renderMask)
 
    // Update lightmap before checking anything that uses alpha
    if (m_lightmap)
-      SetAlpha(100.0f * m_lightmap->m_currentIntensity / (m_lightmap->m_d.m_intensity * m_lightmap->m_d.m_intensity_scale));
+      SetAlpha((int)(100.0f * m_lightmap->m_currentIntensity / (m_lightmap->m_d.m_intensity * m_lightmap->m_d.m_intensity_scale)));
 
    if (m_d.m_color == 0 || m_d.m_alpha == 0.0f || m_d.m_intensity_scale == 0.0f)
       return;
